Fix duplicate helper definitions in extension
