+++ src/dependencies.js
@@ -1,0 +1,3 @@
+export { extension_settings, getContext } from "../../../../extensions.js";
+export { saveSettingsDebounced, event_types, eventSource } from "../../../../../script.js";
+export { executeSlashCommandsOnChatInput, registerSlashCommand } from "../../../../slash-commands.js";
